added hp slider and dmg upgrade

+++ Assets/Scripts/EnemySO.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewEnemy", menuName = "ScriptableObjects/NewEnemy", order = 1)]
 public class EnemySO : ScriptableObject
 {
-    public double hp;
+    public int hp;
     public int goldDrop;
     public Sprite sprite; 
 }
